+++ components/forms/FilterableSearch/react.js
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react'
 import classNames from 'classnames/bind'
 import { mFilter, mChevronDown } from '@masonite/svg-icons'
-import { React as Input } from 'components/forms/Input'
+import { React as TextField } from 'components/forms/TextField'
 import { React as Checkbox } from 'components/forms/Checkbox'
 import PropTypes from 'prop-types'
 import style from './style.module.scss'
@@ -57,7 +57,7 @@ class FilterableSearch extends PureComponent {
       <button
         type="button"
         onClick={() => onSubmit && onSubmit(value)}
-        className={cx(['FilterableSearch__btn'])}
+        className={cx(['filterable-search__btn'])}
       >
         {buttonLabel}
       </button>
@@ -66,7 +66,7 @@ class FilterableSearch extends PureComponent {
     const DropdownToggle = () => (
       <button
         type="button"
-        className={cx(['FilterableSearch__dropdown-toggle'])}
+        className={cx(['filterable-search__dropdown-toggle'])}
         onClick={() =>
           this.setState(prevState => ({
             showFilters: !prevState.showFilters,
@@ -74,7 +74,7 @@ class FilterableSearch extends PureComponent {
         }
       >
         <div
-          className={cx(['FilterableSearch__dropdown-toggle-icon'])}
+          className={cx(['filterable-search__dropdown-toggle-icon'])}
           dangerouslySetInnerHTML={{ __html: selectedDropdownIcon }}
         />
       </button>
@@ -83,15 +83,15 @@ class FilterableSearch extends PureComponent {
     const DropdownMenu = ({ show }) => (
       <ul
         className={cx([
-          'FilterableSearch__dropdown-menu',
-          { 'FilterableSearch__dropdown-menu--open': show },
+          'filterable-search__dropdown-menu',
+          { 'filterable-search__dropdown-menu--open': show },
         ])}
       >
         {filterChoices.map(({ label, value, checked }, index) => (
-          <li className={cx(['FilterableSearch__dropdown-menu-item'])} key={value}>
+          <li className={cx(['filterable-search__dropdown-menu-item'])} key={value}>
             {/* eslint-disable-next-line */}
             <label
-              className={cx(['FilterableSearch__dropdown-menu-item-label'])}
+              className={cx(['filterable-search__dropdown-menu-item-label'])}
               htmlFor={`${value}${index}`}
             >
               <Checkbox
@@ -99,7 +99,7 @@ class FilterableSearch extends PureComponent {
                 checked={checked}
                 onChange={() => onFilterChange({ label, value, checked })}
               />
-              <span className={cx(['FilterableSearch__dropdown-menu-item-label-text'])}>
+              <span className={cx(['filterable-search__dropdown-menu-item-label-text'])}>
                 {label}
               </span>
             </label>
@@ -109,8 +109,8 @@ class FilterableSearch extends PureComponent {
     )
 
     return (
-      <div className={cx(['FilterableSearch'])}>
-        <Input
+      <div className={cx(['filterable-search'])}>
+        <TextField
           placeholder={placeholder}
           onChange={value => onInputChange(value)}
           onKeyDown={ev => {