+++ components/forms/TextField/stories.react.js
@@ -1,20 +1,28 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
+import {
+  withKnobs, boolean, select, text,
+} from '@storybook/addon-knobs'
 import { withState } from '@dump247/storybook-state'
-import { withKnobs, boolean, select } from '@storybook/addon-knobs'
-import Input from './react'
+import TextField from './react'
 
-storiesOf('Forms/Input', module)
+storiesOf('Forms/TextField', module)
   .addDecorator(withKnobs)
   .add(
     'default',
     withState({ value: '' })(({ store }) => {
       const variant = select('Variant', { Light: 'light', Dark: 'dark' }, 'light')
+      const label = text('Label', 'Label')
       const error = boolean('Error State', false)
+      const helper = error ? 'Something is wrong.' : ''
       return (
-        <Input
+        <TextField
           variant={variant}
           error={error}
+          helper={helper}
+          label={label}
+          placeholder="Placeholder"
+          onBlur={value => console.log('onBlur called, value: ', value)}
           onChange={value => store.set({ value })}
           value={store.state.value}
         />