+++ packages/external-ui-react/src/index.js
@@ -1,6 +1,7 @@
 export { React as Checkbox } from 'components/forms/Checkbox'
 export { React as Button } from 'components/forms/button/Button'
-export { React as Input } from 'components/forms/Input'
+export { React as Input, React as TextField } from 'components/forms/TextField'
 export { React as FilterableSearch } from 'components/forms/FilterableSearch'
+export { React as OverflowMenu } from 'components/menu/OverflowMenu'
 export { React as Tabs } from 'components/nav/Tabs'
 export { ReactHeading as Heading, ReactP as P } from 'components/typography'